reorganize folders, remove deprecated handlers
