+++ lib/core/utils/dependencies.dart
@@ -1,7 +1,4 @@
 import 'services.dart';
 
-
 @Deprecated('Use `Services` instead')
 typedef Dependencies = Services;
-
-
