+++ lib/core/utils/loaders/activity_data.dart
@@ -6,8 +6,8 @@ import '../../data/models/activity.dart';
 
 @Deprecated('Migrate to settings in SQL')
 abstract final class ActivityData {
-
   @Deprecated('Migrate to settings in SQL')
+
   /// Loads activities from file.
   factory ActivityData.fromFiles({
     required String raidsPath,
