+++ lib/features/components/buttons/join/join_message_component.dart
@@ -1,12 +1,12 @@
-import '../../core/const/command_exceptions.dart';
-import '../interactor/interactor_component.dart';
+import '../../../../core/const/command_exceptions.dart';
+import '../../../interactor/interactor_component.dart';
 
-class LeaveMessageComponent extends InteractorMessageComponent {
-  const LeaveMessageComponent();
+class JoinMessageComponent extends InteractorMessageComponent {
+  const JoinMessageComponent();
 
   @override
   Future<String> uniqueID(Services services) async {
-    return 'leave';
+    return 'join';
   }
 
   @override
@@ -21,16 +21,16 @@ class LeaveMessageComponent extends InteractorMessageComponent {
     final lfgManager = services.lfgManager;
 
     try {
-      await lfgManager.removeMemberFrom(event.interaction.message!, event.interaction.member!.user!);
-      await event.interaction.respond(MessageBuilder(content: 'Вы покинули LFG'), isEphemeral: true);
+      await lfgManager.addMemberTo(event.interaction.message!, event.interaction.member!.user!);
+      await event.interaction.respond(MessageBuilder(content: 'Вы добавлены в LFG'), isEphemeral: true);
     } on CommandException catch (e) {
       await event.interaction.respond(MessageBuilder(content: e.toHumanMessage()), isEphemeral: true);
     } on Object catch (e, st) {
       await event.interaction.respond(
         MessageBuilder(
-          content: 'Произошла неизвестная ошибка при удалении вас из LFG\n'
+          content: 'Произошла неизвестная ошибка при добавлении вас в LFG\n'
               'Метаданные: $e\n'
-              'Стек вызовов: ${st.toString().substring(0, 200)}',
+              'Стек вызовов: $st',
         ),
         isEphemeral: true,
       );
